requirements.txt
django==2.2.10
djangorestframework==3.11.0
django-object-actions==2.0.0